requirements.txt
fastapi
uvicorn[standard]
pydantic
azureml-core
langchain-openai
langchain
pandas
numpy
scikit-learn
faiss-cpu
python-docx
PyPDF2
tiktoken
httpx
plotly
python-dotenv
python-multipart
logging-config
setuptools<81
flask
flask-login
flask-session
bcrypt
flask-socketio
markdown
